requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.8
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
atomicwrites==1.4.1
attrs==24.2.0
autoray==0.6.12
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
cytoolz==0.12.3
dataclasses==0.6
dataclasses-json==0.6.7
deepai==1.0.1
deprecation==2.1.0
distro==1.9.0
et-xmlfile==1.1.0
Faker==28.0.0
fastapi==0.112.2
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
gotrue==2.9.1
greenlet==3.0.3
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
hyperframe==6.0.1
idna==2.10
importlib-metadata==6.11.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.10
langchain-experimental==0.3.2
langchain-text-splitters==0.3.0
langsmith==0.1.135
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
narwhals==1.9.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.7
packaging==23.2
pandas==2.2.2
Pillow==9.5.0
pluggy==1.5.0
postgrest==0.16.11
postgrest-py==0.4.0
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic-settings==2.5.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pytest==6.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
quimb==1.4.0
realtime==2.0.5
referencing==0.35.1
requests==2.25.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
scipy==1.14.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
storage3==0.7.7
streamlit==1.27.0
StrEnum==0.4.15
supabase==2.7.3
supafunc==0.5.1
tabulate==0.9.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.30.6
validators==0.34.0
watchdog==5.0.3
websockets==13.1
Werkzeug==3.0.4
yarl==1.13.1
zipp==3.20.2
